Remove npm start and explain the problem
So, it appears that the artifact has the generated index.html looking for several scripts that were in the /src/scripts directory but have been consoliodated to a single js file in /dist/resources/. There are also some root directory incorrect assumptions for some json and data files, which are not looking in the /ws4kp/ root properly when relatively referencing via http request. Without solving for these, the deployment will be active but the page won't ever function.